에디터 구조 변경 (user와 동기화)
