+++ src/APP/admin-pages/WritePost/Styled/WritePost.writepost.draft.styles.js
@@ -39,8 +39,9 @@ export const DraftModalTitle = styled.h2`
 export const DraftModalClose = styled.button`
   background-image: url('/img/close.png');
   background-size: cover;
-  width: 1.5rem;
-  height: 1.5rem;
+  background-color: transparent;
+  width: 1rem;
+  height: 1rem;
   border: none;
   cursor: pointer;
 `;
@@ -69,14 +70,16 @@ export const DraftItem = styled.div`
 export const DraftTitle = styled.div`
   ${tokens.typography.B3_M_14};
   color: ${tokens.colors.Grey_8};
-  width: 26rem;
+  width: 26rem;  /* 최대 너비 설정 */
+  white-space: nowrap;  /* 줄바꿈 방지 */
+  overflow: hidden;  /* 넘치는 텍스트 숨김 */
+  text-overflow: ellipsis;  /* 너무 길면 ...으로 표시 */
 `;
 
 export const DraftCategory = styled.div`
   ${tokens.typography.B3_M_14};
   color: ${tokens.colors.Grey_8};
 `;
-
 
 export const DraftDate = styled.div`
   ${tokens.typography.B3_M_14};
