+++ src/APP/admin-pages/WritePost/Styled/WritePost.writepost.main.styles.js
@@ -2,10 +2,6 @@ import styled from 'styled-components';
 import * as tokens from "../../../../tokens";
 
 export const Container = styled.div`
-
-`;
-
-export const MarkdownEditorContainer = styled.div`
   display: flex;
   width: 100%;
   height: 100vh;
