+++ src/APP/admin-pages/WritePost/Styled/WritePost.writepost.preview.styles.js
@@ -3,7 +3,7 @@ import * as tokens from "../../../../tokens";
 
 export const PreviewContainer = styled.div`
   width: 50%;
-  padding: 1.25rem;
+  padding: 2.833rem;
   border: 0.0417rem solid #ffffff;
   background-color: #F5FBFF;
   font-family: 'Pretendard', sans-serif;
