+++ src/APP/admin-pages/WritePost/WritePost.writepost.main.jsx
@@ -7,63 +7,137 @@ import request from '../../Api/request';
 
 export default function WritePost() {
   const location = useLocation();
-  const { id } = useParams();  // 게시글 ID 가져오기
-  const { boardId } = location.state || {}; // 수정 시 전달받은 게시글 ID
-  const [markdownContent, setMarkdownContent] = useState('');
+  const [boardId, setBoardId] = useState(location.state?.boardId || null);
+
   const [title, setTitle] = useState('');
+
   const [categoryCode, setCategoryCode] = useState(null);
   const [category, setCategory] = useState(null);
 
-  useEffect(() => {
-    // 게시글 ID가 있을 경우 수정 데이터를 불러옵니다.
-    if (boardId) {
-      const fetchBoardData = async () => {
-        try {
-          const response = await request.get(`/board/${boardId}`);
-          if (response.isSuccess) {
-            const { title, content, category } = response.result;
-            setTitle(title);
-            setMarkdownContent(content);
-            setCategoryCode(categoryCode); // 카테고리 설정
-            setCategory(category);
-          } else {
-            console.error('게시글 상세 조회 실패:', response.message);
-          }
-        } catch (error) {
-          console.error('게시글 상세 조회 중 오류:', error);
-        }
-      };
+  const [boardFileList, setBoardFileList] = useState([]);
+  const [uploadedImageUrls, setUploadedImageUrls] = useState([]);
 
-      fetchBoardData();
+  const [markdownContent, setMarkdownContent] = useState('');
+
+  const [saveYn, setSaveYn] = useState(location.state?.saveYn);
+
+  
+  // 게시글 상세 조회
+  const fetchBoardData = async () => {
+  try {
+    let response;
+    if (boardId !== null) {
+      if (saveYn == false){ // 임시저장
+        response = await request.get(`/board/draft/${boardId}`);
+      }
+      else { // 수정
+        response = await request.get(`/board/${boardId}`);
+      }
+
+    if (response.isSuccess) {
+      const { title, content, categoryCode, category, boardFileList, saveYn } = response.result;
+      setTitle(title);
+      setMarkdownContent(content);
+      setCategoryCode(categoryCode);
+      setCategory(category);
+      setBoardFileList(boardFileList);
+      setSaveYn(saveYn);
+    } else {
+      console.error('게시글 상세 조회 실패:', response.message);
     }
+  }
+  } catch (error) {
+    console.error('게시글 상세 조회 중 오류:', error);
+  }
+  };
+
+
+  useEffect(() => {
+    fetchBoardData();
   }, [boardId]);
 
+
   useEffect(() => {
-    // 마운트 시 스크롤 비활성화
+    // 스크롤 비활성화
     document.body.style.overflow = 'hidden';
-    
-    // 언마운트 시 스크롤 복원
+  }, []);
+
+  const deleteImageFromS3 = async (fileUrl) => {
+    try {
+      const response = await request.delete('/s3', { params: { fileUrl } });
+      if (!response.isSuccess) {
+        console.error('이미지 삭제 실패:', response.message);
+      }
+    } catch (error) {
+      console.error('이미지 삭제 중 오류:', error);
+    }
+  };
+
+
+  const deleteAllUploadedImages = async () => {
+    const promises = uploadedImageUrls.map((url) => deleteImageFromS3(url));
+    await Promise.all(promises);
+  };
+  
+
+  useEffect(() => {
+    // 페이지를 떠날 때 처리
+    const handleBeforeUnload = (event) => {
+      if (uploadedImageUrls.length > 0) {
+        deleteAllUploadedImages();
+        event.preventDefault();
+        event.returnValue = ''; // 브라우저 기본 메시지 표시
+      }
+    };
+
+    window.addEventListener('beforeunload', handleBeforeUnload);
+
+    // Cleanup: 이벤트 리스너 제거
     return () => {
-      document.body.style.overflow = 'auto';
+      window.removeEventListener('beforeunload', handleBeforeUnload);
+    };
+  }, [uploadedImageUrls]);
+
+  useEffect(() => {
+    // 컴포넌트 언마운트 시 이미지 삭제
+    return () => {
+      deleteAllUploadedImages();
     };
   }, []);
   
   return (
     <Styled.Container>
-      <Styled.MarkdownEditorContainer>
-        <Editor
-          title={title}
-          setTitle={setTitle}
-          setMarkdownContent={setMarkdownContent}
-          initialBoardId={boardId} // 수정 시 boardId 전달
-          initialCategoryCode={categoryCode} // 수정 시 카테고리 코드 전달
-          initialCategory={category}
-          initialContent={markdownContent} // Codemirror 초기 값 전달
-/>
 
-        <Preview title={title} markdownContent={markdownContent} />
+      <Editor
+        boardId={boardId}
+        setBoardId={setBoardId}
 
-      </Styled.MarkdownEditorContainer>
+        fetchBoardData={fetchBoardData}
+
+        title={title}
+        setTitle={setTitle}
+
+        categoryCode={categoryCode}
+        setCategoryCode={setCategoryCode}
+
+        category={category}
+        setCategory={setCategory}
+
+        boardFileList={boardFileList}
+        setBoardFileList={setBoardFileList}
+
+        uploadedImageUrls={uploadedImageUrls}
+        setUploadedImageUrls={setUploadedImageUrls}
+
+        markdownContent={markdownContent}
+        setMarkdownContent={setMarkdownContent}
+
+        saveYn={saveYn}
+        setSaveYn={setSaveYn}
+      />
+
+      <Preview title={title} markdownContent={markdownContent} />
+
     </Styled.Container>
   );
 }
